+++ ext/sqlite3/database.h
@@ -8,6 +8,7 @@ struct _sqlite3Ruby {
     VALUE busy_handler;
     int stmt_timeout;
     struct timespec stmt_deadline;
+    rb_pid_t owner;
 };
 
 typedef struct _sqlite3Ruby sqlite3Ruby;