+++ test/test_resource_cleanup.rb
@@ -17,11 +17,31 @@ module SQLite3
       end
     end
 
+    # # this leaks the result set
     # def test_cleanup_unclosed_resultset_object
     #   db = SQLite3::Database.new(':memory:')
     #   db.execute('create table foo(text BLOB)')
     #   stmt = db.prepare('select * from foo')
     #   stmt.execute
+    # end
+
+    # # this leaks the incompletely-closed connection
+    # def test_cleanup_discarded_connections
+    #   FileUtils.rm_f "test.db"
+    #   db = SQLite3::Database.new("test.db")
+    #   db.execute("create table posts (title text)")
+    #   db.execute("insert into posts (title) values ('hello')")
+    #   db.close
+    #   100.times do
+    #     db = SQLite3::Database.new("test.db")
+    #     db.execute("select * from posts limit 1")
+    #     stmt = db.prepare("select * from posts")
+    #     stmt.execute
+    #     stmt.close
+    #     db.discard
+    #   end
+    # ensure
+    #   FileUtils.rm_f "test.db"
     # end
   end
 end